refactor(file-io): extract HTML-to-Y.Doc conversion from docx.ts
Move htmlToYDoc and 10 DOM traversal helpers into docx-html.ts.
docx.ts keeps loadDocx + exportAnnotations (~120 lines, down from 506).
Re-exports htmlToYDoc for backward compatibility.
